Ore Processing QOLs (#4789)
* Refinery/Smelter UI dust info and click acceptance

* Fixes tool tips, attackbyitem, dust deletion

* Machines with pipes now display location and layer

* All boulder breaking machines drop artifacts

+++ monkestation/code/modules/factory_type_beat/processing/refinery.dm
@@ -12,6 +12,7 @@
 	desc = "BR for short. Accepts boulders and refines non-metallic ores into sheets using internal chemicals. Can be upgraded with stock parts or through chemical inputs."
 	icon_state = "stacker"
 	holds_minerals = TRUE
+	process_string = "Refined Dust"
 	processable_materials = list(
 		/datum/material/glass,
 		/datum/material/plasma,
@@ -42,13 +43,24 @@
 
 /obj/machinery/bouldertech/refinery/add_context(atom/source, list/context, obj/item/held_item, mob/user)
 	. = ..()
-	if(istype(held_item, /obj/item/boulder))
-		context[SCREENTIP_CONTEXT_LMB] = "Insert boulder"
+	if(istype(held_item, /obj/item/boulder) || check_extras(held_item))
+		context[SCREENTIP_CONTEXT_LMB] = "Insert boulder or refined dust"
 	if(istype(held_item, /obj/item/card/id) && points_held > 0)
 		context[SCREENTIP_CONTEXT_LMB] = "Claim mining points"
-	context[SCREENTIP_CONTEXT_RMB] = "Remove boulder"
+	context[SCREENTIP_CONTEXT_RMB] = "Remove boulder or refined dust"
 	return CONTEXTUAL_SCREENTIP_SET
 
+/obj/machinery/bouldertech/refinery/attackby(obj/item/attacking_item, mob/user, params)
+	if(holds_minerals && check_extras(attacking_item)) // Checking for extra items it can refine.
+		var/obj/item/processing/refined_dust/my_dust = attacking_item
+		update_boulder_count()
+		if(!accept_boulder(my_dust))
+			balloon_alert_to_viewers("full!")
+			return
+		balloon_alert_to_viewers("accepted")
+		START_PROCESSING(SSmachines, src)
+		return TRUE
+	return ..()
 
 /**
  * Your other new favorite industrial waste magnet!
@@ -90,7 +102,6 @@
 		set_light_on(TRUE)
 		return TRUE
 
-
 /obj/machinery/bouldertech/refinery/smelter/process()
 	. = ..()
 	if(. == PROCESS_KILL)
@@ -108,12 +119,13 @@
 		if(boulders_concurrent <= 0)
 			break //Try again next time
 
-		if(istype(potential_boulder, /obj/item/processing/refined_dust))
-			refine_dust(potential_boulder)
-
 		if(!istype(potential_boulder, /obj/item/boulder) && !istype(potential_boulder, /obj/item/processing/refined_dust))
 			potential_boulder.forceMove(drop_location())
-			CRASH("\The [src] had a non-boulder in it's boulders contained!")
+			CRASH("\The [src] had a non-boulder in it's boulder container!")
+
+		if(istype(potential_boulder, /obj/item/processing/refined_dust))
+			boulders_concurrent-- //We count dust.
+			refine_dust(potential_boulder)
 
 		var/obj/item/boulder/boulder = potential_boulder
 		if(boulder.durability < 0)
@@ -150,7 +162,15 @@
 	var/obj/item/boulder/disposable_boulder = new (src)
 	disposable_boulder.custom_materials = processable_ores
 	silo_materials.mat_container.insert_item(disposable_boulder, refining_efficiency)
-	qdel(dust)
+	if(length(dust.custom_materials)) // If our dust still has materials we cant process eject it.
+		dust.restart_processing_cooldown() //Reset the cooldown so we don't pick it back up by the same machine.
+		if(isturf(get_step(src, export_side)))
+			dust.forceMove(get_step(src, export_side))
+		else
+			dust.forceMove(drop_location())
+	else
+		qdel(dust)
+
 	playsound(loc, 'sound/weapons/drill.ogg', 50, TRUE, SHORT_RANGE_SOUND_EXTRARANGE)
 	update_boulder_count()
 
@@ -159,8 +179,9 @@
 		return FALSE
 	if(boulders_contained.len >= boulders_held_max)
 		return FALSE
-	if(istype(mover, /obj/item/processing/refined_dust))
-		return TRUE
+	if(check_extras(mover))
+		var/obj/item/processing/refined_dust/dust = mover
+		return dust.can_get_processed()
 	return ..()
 
 /obj/machinery/bouldertech/refinery/return_extras()
